+++ integration_tests/src/tests/mod.rs
@@ -16,8 +16,8 @@ mod evm_test;
 pub use evm_test::*;
 mod extend_sectors_test;
 pub use extend_sectors_test::*;
-// mod init_test;
-// mod market_miner_withdrawal_test;
+mod market_miner_withdrawal_test;
+pub use market_miner_withdrawal_test::*;
 // mod multisig_test;
 // mod power_scenario_tests;
 // mod publish_deals_test;
